Move routes for different versions to individual files
This will make maintaining routing easier in future
